[heroku] try to fix nodegit problems

+++ heroku.sh
@@ -1,6 +1,8 @@
-#!/usr/bin/env bash
+#!/usr/bin/env sh
 
 touch .env
-mv /app/.apt/usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.20 /app/.apt/usr/lib/x86_64-linux-gnu/libstdc++.so.6
+if [ -f  /app/.apt/usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.20 ]; then
+    mv /app/.apt/usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.20 /app/.apt/usr/lib/x86_64-linux-gnu/libstdc++.so.6
+fi
 
 npm start
